[FIX] Unit tests cases

+++ MicroserviceWarehouse/warehouse/src/main/java/com/olvera/warehouse/controller/UserController.java
@@ -95,7 +95,7 @@ public class UserController {
     })
     @GetMapping("/usersProducts")
     public ResponseEntity<PageResponse> getUsersProduct(
-            @RequestParam(value = "userId", defaultValue = "1", required = true)Integer userId,
+            @RequestParam(value = "userId", defaultValue = "1", required = true)Long userId,
             @RequestParam(value = "pageNo", defaultValue = DEFAULT_PAGE_NUMBER, required = false) int pageNo,
             @RequestParam(value = "pageSize", defaultValue = DEFAULT_PAGE_SIZE, required = false) int pageSize
     ) {
